{{ cookiecutter.project_slug }}/api/cars/serializers.py
from rest_framework import serializers

from api.models.car import Car


class CarSerializer(serializers.ModelSerializer):
    class Meta:
        model = Car
        fields = ['id', 'make', 'model', 'year']
